+++ CHANGELOG.md
@@ -14,6 +14,7 @@ i.e. `<Major version>.<Minor version>.<Patch version>`
 - Support enable TLS option for webui
 - Switch config var from ansible to by envar
 - Fix kamailio syslog format
+- Add logging configuration envars
 
 ## [v0.7.0] - 2023-08-027
 - Fix jumps routing table